backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
sqlmodel
psycopg2-binary
redis
google-generativeai
python-dotenv
chromadb
httpx
redis
